fix: finish 'ClaimRow' component refactoring
