+++ src/components/ClaimTile/ClaimTile.scss
@@ -7,5 +7,4 @@
   height: 240px;
   border-radius: 16px;
   border: solid 1px v.$color5;
-
 }
